Create Main class and HandleRentals class
